Fixup: Move urlconf fixture setup into test_relations
