+++ tests/test_relations.py
@@ -1,11 +1,14 @@
 import pytest
+from django.conf.urls import re_path
+from rest_framework.routers import SimpleRouter
 
 from rest_framework_json_api.relations import HyperlinkedRelatedField
+from rest_framework_json_api.views import ModelViewSet, RelationshipView
 
 from .models import BasicModel
 
 
-@pytest.mark.urls("tests.urls")
+@pytest.mark.urls(__name__)
 @pytest.mark.parametrize(
     "format_links,expected_url_segment",
     [
@@ -38,3 +41,34 @@ def test_relationship_urls_respect_format_links(
     actual = field.get_links(model)
 
     assert expected == actual
+
+
+# Routing setup
+
+
+class BasicModelViewSet(ModelViewSet):
+    class Meta:
+        model = BasicModel
+
+
+class BasicModelRelationshipView(RelationshipView):
+    queryset = BasicModel.objects
+
+
+router = SimpleRouter()
+router.register(r"basic_models", BasicModelViewSet, basename="basic-model")
+
+urlpatterns = [
+    re_path(
+        r"^basic_models/(?P<pk>[^/.]+)/(?P<related_field>[^/.]+)/$",
+        BasicModelViewSet.as_view({"get": "retrieve_related"}),
+        name="basic-model-related",
+    ),
+    re_path(
+        r"^basic_models/(?P<pk>[^/.]+)/relationships/(?P<related_field>[^/.]+)/$",
+        BasicModelRelationshipView.as_view(),
+        name="basic-model-relationships",
+    ),
+]
+
+urlpatterns += router.urls
